nocturna-py-back/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
google==3.0.0
google-ai-generativelanguage==0.1.0
google-api-core==2.24.1
google-auth==2.38.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
proto-plus==1.26.0
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.44.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.2.3
uvicorn==0.33.0
zipp==3.20.2
